Move atari registration logic to atari.py
